Move `useQuery` SSR tests to the `ssr` package
Testing SSR functionality in the `hooks` package introduces
a dependency cycle:

@apollo/react-hooks -> @apollo/react-ssr -> @apollo/react-hooks

To avoid this, let's move the `useQuery` SSR tests to the `ssr`
package.
