+++ packages/hooks/package.json
@@ -54,7 +54,6 @@
     "access": "public"
   },
   "devDependencies": {
-    "@apollo/react-ssr": "file:../ssr",
     "jest": "^24.8.0",
     "rollup": "1.16.2",
     "tsc-watch": "^2.2.1",
