+++ packages/hooks/src/__tests__/useQuery.test.tsx
@@ -4,157 +4,54 @@ import gql from 'graphql-tag';
 import { MockedProvider } from '@apollo/react-testing';
 import { render, cleanup } from '@testing-library/react';
 import { useQuery } from '@apollo/react-hooks';
-import { renderToStringWithData } from '@apollo/react-ssr';
 
 describe('useQuery Hook', () => {
   afterEach(cleanup);
 
-  describe('General use', () => {
-    it('should handle a simple query properly', done => {
-      const query: DocumentNode = gql`
-        query {
-          cars {
-            make
-            model
-            vin
-          }
+  it('should handle a simple query properly', done => {
+    const query: DocumentNode = gql`
+      query {
+        cars {
+          make
+          model
+          vin
         }
-      `;
+      }
+    `;
 
-      const resultData = {
-        cars: [
-          {
-            make: 'Audi',
-            model: 'RS8',
-            vin: 'DOLLADOLLABILL',
-            __typename: 'Car'
-          }
-        ]
-      };
-
-      const mocks = [
+    const resultData = {
+      cars: [
         {
-          request: {
-            query
-          },
-          result: { data: resultData }
+          make: 'Audi',
+          model: 'RS8',
+          vin: 'DOLLADOLLABILL',
+          __typename: 'Car'
         }
-      ];
+      ]
+    };
 
-      const Component = () => {
-        const { data, loading } = useQuery(query);
-        if (!loading) {
-          expect(data).toEqual(resultData);
-          done();
-        }
-        return null;
-      };
+    const mocks = [
+      {
+        request: {
+          query
+        },
+        result: { data: resultData }
+      }
+    ];
 
-      render(
-        <MockedProvider mocks={mocks}>
-          <Component />
-        </MockedProvider>
-      );
-    });
-  });
+    const Component = () => {
+      const { data, loading } = useQuery(query);
+      if (!loading) {
+        expect(data).toEqual(resultData);
+        done();
+      }
+      return null;
+    };
 
-  describe('SSR', () => {
-    it('should support SSR', () => {
-      const query: DocumentNode = gql`
-        query {
-          cars {
-            make
-            model
-            vin
-          }
-        }
-      `;
-
-      const resultData = {
-        cars: [
-          {
-            make: 'Audi',
-            model: 'RS8',
-            vin: 'DOLLADOLLABILL',
-            __typename: 'Car'
-          }
-        ]
-      };
-
-      const mocks = [
-        {
-          request: {
-            query
-          },
-          result: { data: resultData }
-        }
-      ];
-
-      const Component = () => {
-        const { data, loading } = useQuery(query);
-        if (!loading) {
-          expect(data).toEqual(resultData);
-          const { make, model, vin } = data.cars[0];
-          return (
-            <div>
-              {make}, {model}, {vin}
-            </div>
-          );
-        }
-        return null;
-      };
-
-      const app = (
-        <MockedProvider mocks={mocks}>
-          <Component />
-        </MockedProvider>
-      );
-
-      return renderToStringWithData(app).then(markup => {
-        expect(markup).toMatch(/Audi/);
-      });
-    });
-
-    it('should initialize data as an empty object when loading', () => {
-      const query: DocumentNode = gql`
-        query {
-          foo {
-            bar
-          }
-        }
-      `;
-
-      const resultData = {
-        foo: {
-          __typename: 'Foo',
-          bar: 'baz'
-        }
-      };
-
-      const mocks = [
-        {
-          request: {
-            query
-          },
-          result: { data: resultData }
-        }
-      ];
-
-      const Component = () => {
-        const { data, loading } = useQuery(query);
-        if (loading) {
-          expect(data).toEqual({});
-        }
-        return null;
-      };
-
-      const app = (
-        <MockedProvider mocks={mocks}>
-          <Component />
-        </MockedProvider>
-      );
-
-      return renderToStringWithData(app);
-    });
+    render(
+      <MockedProvider mocks={mocks}>
+        <Component />
+      </MockedProvider>
+    );
   });
 });
